feat: code organization
